+++ src/Exception/ConnectionException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Exception;
+
+/**
+ * Represents an exception caused by a lost or refused database connection.
+ */
+final class ConnectionException extends Exception {}